refactor: remove user-action dependency
