+++ src/redux/sagas/ga-saga.ts
@@ -1,13 +1,14 @@
 import ReactGA from 'react-ga';
 import { takeLatest } from 'redux-saga/effects';
-import { FIREBASE_USER_RESPONSE, UserFirebaseUserResponseAction } from '../actions/user-action';
+import { userSlice } from '../states/user-state';
+
 
 export default function* gaSaga() {
-  yield takeLatest(FIREBASE_USER_RESPONSE, handleUserAction);
+  yield takeLatest(userSlice.actions.firebaseUserResponse.type, handleUserAction);
 }
 
-function* handleUserAction(action: UserFirebaseUserResponseAction) {
-  const { user } = action;
+function* handleUserAction(action: ReturnType<typeof userSlice.actions.firebaseUserResponse>) {
+  const { payload: user } = action;
   yield ReactGA.set({ userId: user.uid });
   return undefined;
 }
